+++ packages/storage/src/index.ts
@@ -1,0 +1,3 @@
+export * from './adapter.js';
+export * from './local.js';
+export * from './s3.js';